(PC-20410) refacto(libsSearch): move getVenueTypeFacetFilters to libs/algolia
