+++ src/libs/algolia/fetchAlgolia/buildAlgoliaParameters/getVenueTypeFacetFilters.ts
